# server/docs/swagger.yaml
basePath: /api/v1/
definitions:
...
  domain.BasicInformation:
    properties:
      class_type_id:
        type: integer
      code:
        type: string
      description:
        type: string
      grade_id:
        type: integer
      name:
        type: string
      section_id:
        type: integer
      subject_id:
        type: integer
      teacher_id:
        type: integer
    required:
    - class_type_id
    - code
    - description
    - grade_id
    - name
    - section_id
    - subject_id
    - teacher_id
    type: object
  domain.Building:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    required:
    - name
    type: object
  domain.BuildingCreatePayload:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    required:
    - name
    type: object
  domain.BuildingSuccessRes:
    properties:
      data:
        $ref: '#/definitions/domain.Building'
      message:
        type: string
      status:
        type: integer
    type: object
  domain.BuildingUpdatePayload:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    required:
    - name
    type: object
  domain.Class:
    properties:
      basic_info:
        $ref: '#/definitions/domain.BasicInformation'
      created_at:
        type: string
      id:
        type: integer
      schedule_info:
        $ref: '#/definitions/domain.ScheduleInformation'
      updated_at:
        type: string
    required:
    - basic_info
    - schedule_info
    type: object
  domain.ClassCreatePayload:
    properties:
      basic_info:
        $ref: '#/definitions/domain.BasicInformation'
      created_at:
        type: string
      id:
        type: integer
      schedule_info:
        $ref: '#/definitions/domain.ScheduleInformation'
      updated_at:
        type: string
    required:
    - basic_info
    - schedule_info
    type: object
  domain.ClassListRes:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.Class'
        type: array
      message:
        type: string
      meta:
        $ref: '#/definitions/domain.Meta'
      status:
        type: integer
    type: object
  domain.ClassSuccessRes:
    properties:
      data:
        $ref: '#/definitions/domain.Class'
      message:
        type: string
      status:
        type: integer
    type: object
  domain.ClassUpdatePayload:
    properties:
      basic_info:
        $ref: '#/definitions/domain.BasicInformation'
      created_at:
        type: string
      id:
        type: integer
      schedule_info:
        $ref: '#/definitions/domain.ScheduleInformation'
      updated_at:
        type: string
    required:
    - basic_info
    - schedule_info
    type: object
...
  domain.Meta:
    properties:
      current_page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
    type: object
...
  domain.ScheduleInformation:
    properties:
      class_days_ids:
        items:
          type: integer
        type: array
      current_enrollment:
        type: integer
      duration:
        type: integer
      end_date:
        type: string
      end_time:
        type: string
      max_capacity:
        type: integer
      min_capacity:
        type: integer
      room_id:
        type: integer
      start_date:
        type: string
      start_time:
        type: string
    required:
    - class_days_ids
    - duration
    - end_date
    - end_time
    - max_capacity
    - room_id
    - start_date
    - start_time
    type: object
...
  domain.SuccessRes:
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: A comprehensive API for school management system.
  license:
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: School Management System API
  version: "1.0"
paths:
...
  /building/{building_id}:
    delete:
      consumes:
      - application/json
      description: Delete building
      parameters:
      - description: building_id
        in: path
        name: building_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: building
          schema:
            $ref: '#/definitions/domain.SuccessRes'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - JWT: []
      summary: Delete building
      tags:
      - Building
    get:
      consumes:
      - application/json
      description: Get building by id
      parameters:
      - description: building_id
        in: path
        name: building_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: building
          schema:
            $ref: '#/definitions/domain.BuildingSuccessRes'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - JWT: []
      summary: Get building by id
      tags:
      - Building
  /building/create:
    post:
      consumes:
      - application/json
      description: Create building
      parameters:
      - description: building
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/domain.BuildingCreatePayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created new building
          schema:
            $ref: '#/definitions/domain.BuildingSuccessRes'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - JWT: []
      summary: Create building
      tags:
      - Building
  /building/list:
    get:
      consumes:
      - application/json
      description: Get building list
      produces:
      - application/json
      responses:
        "201":
          description: building
          schema:
            $ref: '#/definitions/domain.BuildingSuccessRes'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - JWT: []
      summary: Get building list
      tags:
      - Building
  /building/update:
    put:
      consumes:
      - application/json
      description: Update building
      parameters:
      - description: building
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/domain.BuildingUpdatePayload'
      produces:
      - application/json
      responses:
        "201":
          description: Updated bulding
          schema:
            $ref: '#/definitions/domain.BuildingSuccessRes'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - JWT: []
      summary: Update building
      tags:
      - Building
  /class/{class_id}:
    delete:
      consumes:
      - application/json
      description: Delete class
      parameters:
      - description: class_id
        in: path
        name: class_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: class
          schema:
            $ref: '#/definitions/domain.SuccessRes'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - JWT: []
      summary: Delete class
      tags:
      - Class
    get:
      consumes:
      - application/json
      description: Get class by id
      parameters:
      - description: class_id
        in: path
        name: class_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: class
          schema:
            $ref: '#/definitions/domain.ClassSuccessRes'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - JWT: []
      summary: Get class by id
      tags:
      - Class
  /class/create:
    post:
      consumes:
      - application/json
      description: Create class
      parameters:
      - description: class
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/domain.ClassCreatePayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created new class
          schema:
            $ref: '#/definitions/domain.ClassSuccessRes'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - JWT: []
      summary: Create class
      tags:
      - Class
  /class/list:
    get:
      consumes:
      - application/json
      description: Get class list
      produces:
      - application/json
      responses:
        "201":
          description: class
          schema:
            $ref: '#/definitions/domain.ClassListRes'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - JWT: []
      summary: Get class list
      tags:
      - Class
  /class/update:
    put:
      consumes:
      - application/json
      description: Update class
      parameters:
      - description: class
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/domain.ClassUpdatePayload'
      produces:
      - application/json
      responses:
        "201":
          description: Updated class
          schema:
            $ref: '#/definitions/domain.ClassSuccessRes'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - JWT: []
      summary: Update class
      tags:
      - Class
...
schemes:
- http
- https
securityDefinitions:
  JWT:
    description: JWT security accessToken. Please add it in the format "Bearer {AccessToken}"
      to authorize your requests.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"